refactor(ui): split RecordingDetail.svelte
